+++ python/pod5/src/tests/test_subset.py
@@ -1,15 +1,20 @@
 import json
-import typing
 from pathlib import Path
+from typing import Any, Dict, List, Set, Tuple
 
 import jsonschema
 import pytest
 
+import pod5 as p5
+from pod5.tools.pod5_inspect import do_reads_command
 from pod5.tools.pod5_subset import (
-    assert_no_duplicate_reads,
-    assert_no_missing_reads,
+    calculate_transfers,
+    create_default_filename_template,
+    launch_subsetting,
     parse_csv_mapping,
     parse_json_mapping,
+    parse_table_mapping,
+    resolve_targets,
 )
 
 CSV_CONTENT_1 = """
@@ -24,6 +29,207 @@ CSV_RESULT_1 = {
     "handle_spaces": {"r2", "r3", "r5"},
 }
 
+TEST_DATA_PATH = Path(__file__).parent.parent.parent.parent.parent / "test_data"
+POD5_PATH = TEST_DATA_PATH / "multi_fast5_zip_v3.pod5"
+
+
+class TestSubset:
+    """Test that pod5 subset subsets files"""
+
+    def test_subset_base(self, tmp_path: Path):
+        """Test a known-good basic use case"""
+        # Known good mapping
+        mapping = {
+            "well-2.pod5": {
+                "002fde30-9e23-4125-9eae-d112c18a81a7",
+            },
+            "well-4.pod5": {
+                "00919556-e519-4960-8aa5-c2dfa020980c",
+                "0000173c-bf67-44e7-9a9c-1ad0bc728e74",
+                "00925f34-6baf-47fc-b40c-22591e27fb5c",
+            },
+        }
+
+        resolved_targets = resolve_targets(tmp_path, mapping)
+        transfers = calculate_transfers([POD5_PATH], resolved_targets, False, False)
+        launch_subsetting(transfers)
+
+        # Assert only the expected files are output
+        expected_outnames = list(mapping.keys())
+        actual_outnames = list(path.name for path in tmp_path.glob("*.pod5"))
+        assert sorted(expected_outnames) == sorted(actual_outnames)
+
+        # Check all read_ids are present in their respective files
+        for outname in expected_outnames:
+            with p5.Reader(tmp_path / outname) as reader:
+                assert sorted(reader.read_ids) == sorted(list(mapping[outname]))
+
+    def test_subset_shared_read_id(self, tmp_path: Path):
+        """Test subsample with a mapping with duplicates a read_id"""
+        # Known good mapping
+        mapping = {
+            "well-2.pod5": {
+                "002fde30-9e23-4125-9eae-d112c18a81a7",
+                "00925f34-6baf-47fc-b40c-22591e27fb5c",
+            },
+            "well-4.pod5": {
+                "00919556-e519-4960-8aa5-c2dfa020980c",
+                "0000173c-bf67-44e7-9a9c-1ad0bc728e74",
+                "00925f34-6baf-47fc-b40c-22591e27fb5c",
+            },
+        }
+
+        resolved_targets = resolve_targets(tmp_path, mapping)
+        transfers = calculate_transfers([POD5_PATH], resolved_targets, False, False)
+        launch_subsetting(transfers)
+
+        # Assert only the expected files are output
+        expected_outnames = list(mapping.keys())
+        actual_outnames = list(path.name for path in tmp_path.glob("*.pod5"))
+        assert sorted(expected_outnames) == sorted(actual_outnames)
+
+        # Check all read_ids are present in their respective files
+        for outname in expected_outnames:
+            with p5.Reader(tmp_path / outname) as reader:
+                assert sorted(reader.read_ids) == sorted(list(mapping[outname]))
+
+
+class TestSubsetFilenameTemplating:
+    """Test the output filename templating"""
+
+    @pytest.mark.parametrize(
+        "columns,expected",
+        [
+            (["mux"], "mux-{mux}.pod5"),
+            (["mux", "channel"], "mux-{mux}_channel-{channel}.pod5"),
+            (["channel", "mux"], "channel-{channel}_mux-{mux}.pod5"),
+        ],
+    )
+    def test_default_template(self, columns: List[str], expected: str):
+        template = create_default_filename_template(columns)
+        assert template == expected
+
+
+class TestTableMappingParser:
+    """Test the table parsing functionality"""
+
+    @staticmethod
+    def _create_inspect_reads_mapping(
+        tmp_path: Path, capsys: pytest.CaptureFixture, columns: List[str]
+    ) -> Tuple[Dict, Dict]:
+        # Run pod5 inspect reads
+        with p5.Reader(POD5_PATH) as reader:
+            do_reads_command(reader)
+
+        # Capture stdout from pod5 inspect reads
+        captured_stdout = str(capsys.readouterr().out)
+
+        # CSV
+        csv_path = tmp_path / "table.csv"
+        with csv_path.open("w") as csv:
+            csv.writelines(captured_stdout.splitlines(keepends=True))
+
+        csv_table_mapping = parse_table_mapping(
+            csv_path,
+            filename_template=create_default_filename_template(columns),
+            subset_columns=columns,
+        )
+
+        # TSV
+        tsv_path = tmp_path / "table.tsv"
+        with tsv_path.open("w") as csv:
+            tsv = captured_stdout.replace(",", "\t")
+            csv.writelines(tsv.splitlines(keepends=True))
+
+        tsv_table_mapping = parse_table_mapping(
+            tsv_path,
+            filename_template=create_default_filename_template(columns),
+            subset_columns=columns,
+        )
+
+        return csv_table_mapping, tsv_table_mapping
+
+    def test_table_parsing_well(self, tmp_path: Path, capsys: pytest.CaptureFixture):
+        """Use pod5 inspect reads to produce a well subsampling"""
+
+        columns = ["well"]
+        csv, tsv = self._create_inspect_reads_mapping(tmp_path, capsys, columns)
+
+        # Manually checked
+        expected_mapping = {
+            "well-2.pod5": {
+                "002fde30-9e23-4125-9eae-d112c18a81a7",
+                "009dc9bd-c5f4-487b-ba4c-b9ce7e3a711e",
+                "008468c3-e477-46c4-a6e2-7d021a4ebf0b",
+                "00728efb-2120-4224-87d8-580fbb0bd4b2",
+                "007cc97e-6de2-4ff6-a0fd-1c1eca816425",
+            },
+            "well-4.pod5": {
+                "00919556-e519-4960-8aa5-c2dfa020980c",
+                "0000173c-bf67-44e7-9a9c-1ad0bc728e74",
+                "008ed3dc-86c2-452f-b107-6877a473d177",
+                "006d1319-2877-4b34-85df-34de7250a47b",
+                "00925f34-6baf-47fc-b40c-22591e27fb5c",
+            },
+        }
+
+        assert csv == expected_mapping
+        assert tsv == expected_mapping
+
+    def test_csv_table_parsing_channel(
+        self, tmp_path: Path, capsys: pytest.CaptureFixture
+    ):
+        """Use pod5 inspect reads test channel subsampling"""
+
+        columns = ["channel"]
+        csv, tsv = self._create_inspect_reads_mapping(tmp_path, capsys, columns)
+
+        # Manually checked
+        expected_mapping = {
+            "channel-109.pod5": {"0000173c-bf67-44e7-9a9c-1ad0bc728e74"},
+            "channel-126.pod5": {"007cc97e-6de2-4ff6-a0fd-1c1eca816425"},
+            "channel-147.pod5": {"00728efb-2120-4224-87d8-580fbb0bd4b2"},
+            "channel-199.pod5": {"00919556-e519-4960-8aa5-c2dfa020980c"},
+            "channel-2.pod5": {"008468c3-e477-46c4-a6e2-7d021a4ebf0b"},
+            "channel-452.pod5": {"009dc9bd-c5f4-487b-ba4c-b9ce7e3a711e"},
+            "channel-463.pod5": {"002fde30-9e23-4125-9eae-d112c18a81a7"},
+            "channel-474.pod5": {"008ed3dc-86c2-452f-b107-6877a473d177"},
+            "channel-489.pod5": {"006d1319-2877-4b34-85df-34de7250a47b"},
+            "channel-53.pod5": {"00925f34-6baf-47fc-b40c-22591e27fb5c"},
+        }
+
+        assert csv == expected_mapping
+        assert tsv == expected_mapping
+
+    def test_csv_table_parsing_mixed(
+        self, tmp_path: Path, capsys: pytest.CaptureFixture
+    ):
+        """Use pod5 inspect reads test channel / end_reason (text) subsampling"""
+
+        columns = ["well", "end_reason"]
+        csv, tsv = self._create_inspect_reads_mapping(tmp_path, capsys, columns)
+
+        # Manually checked
+        expected_mapping = {
+            "well-2_end_reason-unknown.pod5": {
+                "002fde30-9e23-4125-9eae-d112c18a81a7",
+                "009dc9bd-c5f4-487b-ba4c-b9ce7e3a711e",
+                "008468c3-e477-46c4-a6e2-7d021a4ebf0b",
+                "00728efb-2120-4224-87d8-580fbb0bd4b2",
+                "007cc97e-6de2-4ff6-a0fd-1c1eca816425",
+            },
+            "well-4_end_reason-unknown.pod5": {
+                "00919556-e519-4960-8aa5-c2dfa020980c",
+                "0000173c-bf67-44e7-9a9c-1ad0bc728e74",
+                "008ed3dc-86c2-452f-b107-6877a473d177",
+                "006d1319-2877-4b34-85df-34de7250a47b",
+                "00925f34-6baf-47fc-b40c-22591e27fb5c",
+            },
+        }
+
+        assert csv == expected_mapping
+        assert tsv == expected_mapping
+
 
 class TestCSVMappingParser:
     """Test the CSV Mapping functionality"""
@@ -33,7 +239,7 @@ class TestCSVMappingParser:
         self,
         tmp_path: Path,
         csv_content: str,
-        result: typing.Dict[str, typing.Set[str]],
+        result: Dict[str, Set[str]],
     ):
         """
         Given an example csv input mapping, parse it and assert it's content
@@ -50,9 +256,7 @@ class TestJSONMappingParser:
     """Test the JSON Mapping functionality"""
 
     @classmethod
-    def _write_example_json_mapping(
-        cls, output: Path, json_mapping: typing.Dict[typing.Any, typing.Any]
-    ):
+    def _write_example_json_mapping(cls, output: Path, json_mapping: Dict[Any, Any]):
         """
         Create an example json subset mapping file in output using the supplied json_mapping
         """
@@ -77,7 +281,7 @@ class TestJSONMappingParser:
         ],
     )
     def test_good_parse_json_mapping(
-        self, tmp_path: Path, json_content: typing.Dict[typing.Any, typing.Any]
+        self, tmp_path: Path, json_content: Dict[Any, Any]
     ):
         """
         Test that given a json mapping file that the parser correctly identifies if it
@@ -90,7 +294,7 @@ class TestJSONMappingParser:
         assert parsed_json == json_content
 
     @pytest.mark.parametrize("bad_mapping", ["read_id", [], {}, [""], ["ok", []]])
-    def test_bad_parse_json_mapping(self, tmp_path: Path, bad_mapping: typing.Any):
+    def test_bad_parse_json_mapping(self, tmp_path: Path, bad_mapping: Any):
         """
         Test that given a json mapping that is badly formatted that the parser
         correctly identifies if it as such
@@ -99,53 +303,3 @@ class TestJSONMappingParser:
         with pytest.raises(jsonschema.exceptions.ValidationError):
             self._write_example_json_mapping(example_json, {"target": bad_mapping})
             parse_json_mapping(example_json)
-
-
-class TestSubsetAssertions:
-    """Test the runtime assertions in the subset application"""
-
-    @pytest.mark.parametrize(
-        "selection,transfers",
-        [({"r1", "r2"}, {"": {"r1", "r2"}})],
-    )
-    def test_passing_no_missing_reads(self, selection, transfers):
-        """Test that assert_no_missing_reads correctly detects all reads"""
-        assert_no_missing_reads(selection, transfers)
-
-    @pytest.mark.parametrize(
-        "selection,transfers,missing",
-        [({"r1", "r2"}, {"": {"r1"}}, 1), ({"r1", "r2", "r3"}, {"": {}}, 3)],
-    )
-    def test_failing_no_missing_reads(self, selection, transfers, missing):
-        """Test that assert_no_missing_reads correctly detects missing reads"""
-        with pytest.raises(AssertionError) as exc:
-            assert_no_missing_reads(selection, transfers)
-
-        expected_err = f"Missing {missing} read_ids from input but --missing_ok not set"
-        assert str(exc.value) == expected_err
-
-    @pytest.mark.parametrize(
-        "mapping",
-        [
-            {"a": {"r1", "r2"}},
-            {"a": {"r1", "r2"}, "b": {"r3"}},
-        ],
-    )
-    def test_passing_no_duplicate_reads(self, mapping) -> None:
-        """Test that assert_no_duplicate_reads correctly detect no duplicates"""
-        assert assert_no_duplicate_reads(mapping) is None  # type: ignore [func-returns-value]
-
-    @pytest.mark.parametrize(
-        "mapping",
-        [
-            {"a": {"r1", "r2"}, "b": {"r1"}},
-            {"a": {"r1", "r2"}, "b": {"r3", "r2"}},
-        ],
-    )
-    def test_failing_no_duplicate_reads(self, mapping):
-        """Test that assert_no_duplicate_reads correctly detect duplicates"""
-        with pytest.raises(AssertionError) as exc:
-            assert_no_duplicate_reads(mapping)
-
-        expected_err = "Duplicate outputs detected but --duplicate_ok not set"
-        assert str(exc.value) == expected_err